Fix bug with duplicate stint entries on the race start after qualy.
Issue was that realtime data and shared memory data are not in sync, shared mem updates faster and thus Session.isNewSession lags behind a bit.
Fix however could be to always use Helpers.RaceSessionTypeFromString or use the fact that real new stint can only start if we don't jump to menu screen.