fix: input tyre pressures not being set correctly at race start
Potentially fixed a larger issue but in AC at race start the pressures are always visible. When we load the game, we set the pressures. When one then changes the pressures in the setup menu, it registers as a pressure loss.

This swaps the branches. If we are moving, we cannot be getting a new input pressure. Also when we are standing we cannot be getting a pressure losses.