fix: ideal tyre temps and pressures were not updated from default
The way we exported the ideal values relied on the tyre info being the same object. This however was not the case as a new object is assigned to it after every tyre change.

We need to follow the references all the time to reach correct tyre info.

For future improvements we could potentially overwrite the delegates after the tyre change, since for the duration of the stint these will always point to the same location.